+++ mashkin.andrew/I2/maxCountOfEqualElements.cpp
@@ -1,9 +1,9 @@
 #include "maxCountOfEqualElements.h"
 #include <cstddef>
 
-void countMaxQuantityOfEqualElements(const int* arr, const size_t size, size_t& count, size_t& maxCount)
+void countMaxQuantityOfEqualElements(const int* arr, const int size, size_t& count, size_t& maxCount)
 {
-  for (size_t i = 0; i < size - 1; ++i)
+  for (int i = 0; i < size - 1; ++i)
   {
     if (arr[i] == arr[i + 1])
     {
