+++ mashkin.andrew/I2/samePositiveElements.cpp
@@ -1,15 +1,15 @@
 #include "samePositiveElements.h"
 #include <cstddef>
 
-bool isThereTheSamePositiveElements(const int* arr, const size_t size, bool& ans)
+bool isThereTheSamePositiveElements(const int* arr, const int size, bool& ans)
 {
-  for (size_t i = 0; i + 1 < size; ++i)
+  for (int i = 0; i + 1 < size; ++i)
   {
     if (arr[i] < 1)
     {
       continue;
     }
-    for (size_t j = i + 1; j < size; ++j)
+    for (int j = i + 1; j < size; ++j)
     {
       if (arr[j] < 1)
       {
