+++ mashkin.andrew/I2/samePositiveElements.h
@@ -2,5 +2,5 @@
 #define SAMEPOSITIVEELEMENTS_H
 #include <cstddef>
 
-bool isThereTheSamePositiveElements(const int* arr, const size_t size, bool& ans);
+bool isThereTheSamePositiveElements(const int* arr, const int size, bool& ans);
 #endif
